+++ packages/terra-core-docs/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Added
   * Added examples and tests for `terra-dropdown-button` `SplitButton` with icons.
 
+* Changed
+  * Updated documentation for `terra-signature`.
+
 ## 1.69.0 - (March 27, 2024)
 
 * Added